ReadFile
This script compares the last scanned QR Code with a list of names and file location stored in a txt files. When it finds a match it plays back a corresponding mp3 file. The txt file has the following format:
qrcode1, "c:\somefolder\file1.mp3"
qecode2, "c:\somefolder\file2.mp3"